service endpoints and handlers

+++ FadeAfro/FadeAfro.Application/Features/Services/DeleteService/DeleteServiceHandler.cs
@@ -1,3 +1,4 @@
+using FadeAfro.Domain.Exceptions.MasterProfile;
 using FadeAfro.Domain.Exceptions.Service;
 using FadeAfro.Domain.Repositories;
 using MediatR;
@@ -7,18 +8,29 @@ namespace FadeAfro.Application.Features.Services.DeleteService;
 public class DeleteServiceHandler : IRequestHandler<DeleteServiceCommand, Unit>
 {
     private readonly IServiceRepository _serviceRepository;
-
-    public DeleteServiceHandler(IServiceRepository serviceRepository)
+    private readonly IMasterProfileRepository _masterProfileRepository;
+    
+    public DeleteServiceHandler(
+        IServiceRepository serviceRepository,
+        IMasterProfileRepository masterProfileRepository)
     {
         _serviceRepository = serviceRepository;
+        _masterProfileRepository = masterProfileRepository;
     }
 
     public async Task<Unit> Handle(DeleteServiceCommand command, CancellationToken cancellationToken)
     {
+        var masterProfile = await _masterProfileRepository.GetByMasterIdAsync(command.UserId);
+        if (masterProfile is null)
+            throw new MasterProfileNotFoundException();
+        
         var service = await _serviceRepository.GetByIdAsync(command.ServiceId);
 
         if (service is null)
             throw new ServiceNotFoundException();
+        
+        if (service.MasterProfileId != masterProfile.Id)
+            throw new ServiceFromAnotherMasterException();
 
         await _serviceRepository.DeleteAsync(command.ServiceId);
